src/pyproject.toml
[project]
name = "mcp_zabbix"
version = "0.1.0"
description = "mcp zabbix"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.9.1",
    "httpx>=0.28.1",
    "load-dotenv>=0.1.0",
    "mcp[cli]>=1.10.1",
]
